// src/lib/types.ts
export type Token = {
    name: string;
    symbol: string;
    decimals: number;
    address: string;
  };
  

  export type Partner = {
    id: string;
    website: string;
    logo: string;
    name: string;
    twapId: string;
    liquidityHubChains: number[];
    liquidityHubID: string
  };
  